design: 나의서재 별점/한줄평, 독후감 모달 스타일 추가
