+++ client/src/components/molecules/MyBooks/Books/Book/styles.ts
@@ -1,25 +1,14 @@
 import styled from 'styled-components';
-import { THEME } from '../../../../constants';
-
-/** 나의 서재 전체 도서 목록 컨테이너 section 태그입니다. */
-export const BookCards = styled.section`
-  margin: 0 auto;
-  max-width: 1000px;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 20px;
-
-  @media screen and (max-width: 479px) {
-    display: flex;
-    flex-direction: column;
-  }
-`;
+import { THEME } from '../../../../../constants';
 
 /** 개별 도서 테두리 영역 article 태그입니다. */
 export const BookCard = styled.article`
   box-shadow: 5px 5px 5px rgba(0, 0, 0, 0.1);
-  padding: 1rem;
   display: flex;
+  border-radius: ${THEME.RADIUS};
+  width: 250px;
+  height: 450px;
+
   flex-direction: column;
   justify-content: center;
   align-items: center;
@@ -27,39 +16,18 @@ export const BookCard = styled.article`
   cursor: pointer;
 `;
 
-/** 개별 도서 책 표지 img 태그입니다. */
-export const BookImg = styled.img`
-  width: 80%;
-
-  @media screen and (max-width: 479px) {
-    width: max-content;
-  }
-`;
-
 /** 개별 도서의 정보 표시영역 div 태그입니다 */
-export const BookInfo = styled.div`
+export const Info = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  width: 150px;
-  height: 150px;
   padding: 0 0.5rem;
+  height: 70%;
 
   @media screen and (max-width: 479px) {
     /* width: max-content; */
   }
-`;
-
-/** 개별 도서 책 제목 h4 태그입니다. */
-export const BookTitle = styled.h3`
-  padding: 0.5rem;
-`;
-
-/** 개별 도서 별점영역을 감싸는 평점 추가 button 태그입니다. */
-export const BookRate = styled.button`
-  cursor: pointer;
-  border: none;
 `;
 
 /** 개별 도서의 컨텐츠 영역 div 태그입니다. */
@@ -70,9 +38,24 @@ export const BookContent = styled.div`
   padding: 15px;
 `;
 
+export const ReviewTitle = styled.h2`
+  font-size: 24px;
+  padding: 1rem 0.5rem;
+  color: ${THEME.STRONG};
+`;
+
+export const ReviewView = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
 export const ReportView = styled.div`
   width: 100%;
   display: flex;
+  gap: 1rem;
   flex-direction: column;
   justify-content: center;
   align-items: center;
